requirements.txt
# Minimal requirements for Alana Legal Sense - Production Ready
# Core Web Framework
Flask==3.1.2
gunicorn==23.0.0
flask-cors

# AI and Document Processing
openai==2.8.1
langchain-openai==1.0.3
langchain-core==1.0.6
langchain-community==0.4.1
langchain-text-splitters==1.0.0
faiss-cpu==1.13.0

# Machine Learning (CRITICAL - needed for chatbot.model)
scikit-learn==1.7.2
numpy==2.3.4
pandas==2.3.3

# Document Processing
pypdf==6.3.0
python-docx==1.2.0

# Database and Storage  
supabase==2.24.0

# Essential Dependencies
python-dotenv==1.2.1
requests==2.32.5

# LangChain dependencies
pydantic==2.12.4
tenacity==9.1.2
PyYAML==6.0.3
jsonschema==4.25.1

# Security
cryptography==46.0.3

# Essential Flask deps
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0

# Core Python Support
packaging==25.0
typing_extensions==4.15.0
certifi==2025.11.12
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11